fix active state of Chip component in Filter and Search
