+++ src/contexts/RestaurantsContext.tsx
@@ -2,7 +2,7 @@ import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { FC, PropsWithChildren, createContext, useCallback, useEffect, useState } from 'react';
 import i18n from '../i18n';
 import { Restaurant, restaurantsService } from '../utils/api/restaurantsService/restaurantsService';
-import { mockMeals } from '../pages/Restaurants/MockRestaurantsList';
+import { options, types } from '../pages/Restaurants/MockRestaurantsList';
 
 export type Option = {
     /**
@@ -207,26 +207,6 @@ export const RestaurantsProvider: FC<PropsWithChildren> = ({ children }) => {
         },
         [queryClient]
     );
-    const increment = (function (n) {
-        return function () {
-            n += 1;
-            return n;
-        };
-    })(0);
-    const options = mockMeals
-        .map(({ meals, restaurantName }) => {
-            return meals.map((meal) => {
-                return [meal.name, restaurantName];
-            });
-        })
-        .flat(2)
-        .filter((option, i, ar) => ar.indexOf(option) === i)
-        .map((option) => {
-            return { id: increment(), name: option };
-        });
-    const types = ['fastFood', 'cafe', 'cafeBar'].map((type) => {
-        return { id: increment(), name: type, selected: false };
-    });
     const addOption = (option: Option) => {
         if (!selectedOptions.find((opt: Option) => opt.id === option.id)) {
             setSelectedOptions([...selectedOptions, option]);
